requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0